Update jquery.history.js
I produce a set of documentation templates that need to work just as well without a web server and I'm using history.js for this in html4 mode when there's no server. I had to make this change to the History.createStateObject method for IE, otherwise it gets into a loop appending copies of the URL onto itself. I have only modified the jQuery version because that's the one I'm using. I can imagine that the problem also exists in the other versions, however.